api/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
psycopg2-binary

